senses/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
billiard==3.6.4.0
boto3==1.24.61
botocore==1.27.61
celery==5.2.7
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
jmespath==1.0.1
kombu==5.2.4
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.30
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
